requirements.txt
streamlit
langchain
python-dotenv
langchain_groq
langchain_astradb
sentence-transformers
watchdog